Don't depend on cc_library bar from main_import
This removes spurious linker flags that try to link against static libraries of bar and its transitive dependencies (e.g. `bazel-out/darwin_arm64-fastbuild/bin/cc/lib/libbar.a`).

It's fine for the main_dynamic_deps target because with the combination of cc_shared_library and dynamic_deps, Bazel knows which shared library was compiled with which cc_library, and it only links to static libraries that are not part of shared libraries it depends on.